Move tests to the common part where they belong
