+++ src/common/component.rs
@@ -292,3 +292,15 @@ impl Component {
         self.inner.refresh()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::*;
+
+    #[test]
+    fn test_components_mac_m1() {
+        let mut components = Components::new();
+        components.refresh_list();
+        components.refresh_list();
+    }
+}
