+++ src/lib.rs
@@ -184,7 +184,6 @@ mod doctest {
 #[cfg(test)]
 mod test {
     use crate::*;
-    use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 
     #[cfg(feature = "unknown-ci")]
     #[test]
@@ -202,111 +201,6 @@ mod test {
         check_is_supported(IS_SUPPORTED_SYSTEM);
         check_supported_signals(SUPPORTED_SIGNALS);
         check_minimum_cpu_update_interval(MINIMUM_CPU_UPDATE_INTERVAL);
-    }
-
-    #[test]
-    fn check_process_memory_usage() {
-        let mut s = System::new();
-        s.refresh_specifics(RefreshKind::everything());
-
-        if IS_SUPPORTED_SYSTEM {
-            // No process should have 0 as memory usage.
-            #[cfg(not(feature = "apple-sandbox"))]
-            assert!(!s.processes().iter().all(|(_, proc_)| proc_.memory() == 0));
-        } else {
-            // There should be no process, but if there is one, its memory usage should be 0.
-            assert!(s.processes().iter().all(|(_, proc_)| proc_.memory() == 0));
-        }
-    }
-
-    #[test]
-    fn check_system_implemented_traits() {
-        fn check<T: Sized + std::fmt::Debug + Default + Send + Sync>(_: T) {}
-
-        check(System::new());
-    }
-
-    #[test]
-    fn check_memory_usage() {
-        let mut s = System::new();
-
-        assert_eq!(s.total_memory(), 0);
-        assert_eq!(s.free_memory(), 0);
-        assert_eq!(s.available_memory(), 0);
-        assert_eq!(s.used_memory(), 0);
-        assert_eq!(s.total_swap(), 0);
-        assert_eq!(s.free_swap(), 0);
-        assert_eq!(s.used_swap(), 0);
-
-        s.refresh_memory();
-        if IS_SUPPORTED_SYSTEM {
-            assert!(s.total_memory() > 0);
-            assert!(s.used_memory() > 0);
-            if s.total_swap() > 0 {
-                // I think it's pretty safe to assume that there is still some swap left...
-                assert!(s.free_swap() > 0);
-            }
-        } else {
-            assert_eq!(s.total_memory(), 0);
-            assert_eq!(s.used_memory(), 0);
-            assert_eq!(s.total_swap(), 0);
-            assert_eq!(s.free_swap(), 0);
-        }
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn check_processes_cpu_usage() {
-        if !IS_SUPPORTED_SYSTEM {
-            return;
-        }
-        let mut s = System::new();
-
-        s.refresh_processes();
-        // All CPU usage will start at zero until the second refresh
-        assert!(s
-            .processes()
-            .iter()
-            .all(|(_, proc_)| proc_.cpu_usage() == 0.0));
-
-        // Wait a bit to update CPU usage values
-        std::thread::sleep(MINIMUM_CPU_UPDATE_INTERVAL);
-        s.refresh_processes();
-        assert!(s
-            .processes()
-            .iter()
-            .all(|(_, proc_)| proc_.cpu_usage() >= 0.0
-                && proc_.cpu_usage() <= (s.cpus().len() as f32) * 100.0));
-        assert!(s
-            .processes()
-            .iter()
-            .any(|(_, proc_)| proc_.cpu_usage() > 0.0));
-    }
-
-    #[test]
-    fn check_cpu_usage() {
-        if !IS_SUPPORTED_SYSTEM {
-            return;
-        }
-        let mut s = System::new();
-        for _ in 0..10 {
-            s.refresh_cpu_usage();
-            // Wait a bit to update CPU usage values
-            std::thread::sleep(MINIMUM_CPU_UPDATE_INTERVAL);
-            if s.cpus().iter().any(|c| c.cpu_usage() > 0.0) {
-                // All good!
-                return;
-            }
-        }
-        panic!("CPU usage is always zero...");
-    }
-
-    #[test]
-    fn check_list() {
-        let mut users = Users::new();
-        assert!(users.list().is_empty());
-        users.refresh_list();
-        assert!(users.list().len() >= MIN_USERS);
     }
 
     #[test]
@@ -372,98 +266,12 @@ mod test {
     }
 
     #[test]
-    fn check_system_info() {
-        // We don't want to test on unsupported systems.
-        if IS_SUPPORTED_SYSTEM {
-            assert!(!System::name()
-                .expect("Failed to get system name")
-                .is_empty());
-
-            assert!(!System::kernel_version()
-                .expect("Failed to get kernel version")
-                .is_empty());
-
-            assert!(!System::os_version()
-                .expect("Failed to get os version")
-                .is_empty());
-
-            assert!(!System::long_os_version()
-                .expect("Failed to get long OS version")
-                .is_empty());
-        }
-
-        assert!(!System::distribution_id().is_empty());
-    }
-
-    #[test]
-    fn check_host_name() {
-        // We don't want to test on unsupported systems.
-        if IS_SUPPORTED_SYSTEM {
-            assert!(System::host_name().is_some());
-        }
-    }
-
-    #[test]
-    fn check_refresh_process_return_value() {
-        // We don't want to test on unsupported systems.
-        if IS_SUPPORTED_SYSTEM {
-            let _pid = get_current_pid().expect("Failed to get current PID");
-
-            #[cfg(not(feature = "apple-sandbox"))]
-            {
-                let mut s = System::new();
-                // First check what happens in case the process isn't already in our process list.
-                assert!(s.refresh_process(_pid));
-                // Then check that it still returns true if the process is already in our process list.
-                assert!(s.refresh_process(_pid));
-            }
-        }
-    }
-
-    #[test]
-    fn test_components_mac_m1() {
-        let mut components = Components::new();
-        components.refresh_list();
-        components.refresh_list();
-    }
-
-    #[test]
     fn ensure_is_supported_is_set_correctly() {
         if MIN_USERS > 0 {
             assert!(IS_SUPPORTED_SYSTEM);
         } else {
             assert!(!IS_SUPPORTED_SYSTEM);
         }
-    }
-
-    #[test]
-    fn check_cpus_number() {
-        let mut s = System::new();
-
-        // This information isn't retrieved by default.
-        assert!(s.cpus().is_empty());
-        if IS_SUPPORTED_SYSTEM {
-            // The physical cores count is recomputed every time the function is called, so the
-            // information must be relevant even with nothing initialized.
-            let physical_cores_count = s
-                .physical_core_count()
-                .expect("failed to get number of physical cores");
-
-            s.refresh_cpu_usage();
-            // The cpus shouldn't be empty anymore.
-            assert!(!s.cpus().is_empty());
-
-            // In case we are running inside a VM, it's possible to not have a physical core, only
-            // logical ones, which is why we don't test `physical_cores_count > 0`.
-            let physical_cores_count2 = s
-                .physical_core_count()
-                .expect("failed to get number of physical cores");
-            assert!(physical_cores_count2 <= s.cpus().len());
-            assert_eq!(physical_cores_count, physical_cores_count2);
-        } else {
-            assert_eq!(s.physical_core_count(), None);
-        }
-        assert!(s.physical_core_count().unwrap_or(0) <= s.cpus().len());
     }
 
     #[test]
@@ -480,160 +288,5 @@ mod test {
                 "SUPPORTED_SIGNALS should be empty on not support systems!"
             );
         }
-    }
-
-    // Ensure that the CPUs frequency isn't retrieved until we ask for it.
-    #[test]
-    fn check_cpu_frequency() {
-        if !IS_SUPPORTED_SYSTEM {
-            return;
-        }
-        let mut s = System::new();
-        s.refresh_processes();
-        for proc_ in s.cpus() {
-            assert_eq!(proc_.frequency(), 0);
-        }
-        s.refresh_cpu_usage();
-        for proc_ in s.cpus() {
-            assert_eq!(proc_.frequency(), 0);
-        }
-        // In a VM, it'll fail.
-        if std::env::var("APPLE_CI").is_err() && std::env::var("FREEBSD_CI").is_err() {
-            s.refresh_cpu_specifics(CpuRefreshKind::everything());
-            for proc_ in s.cpus() {
-                assert_ne!(proc_.frequency(), 0);
-            }
-        }
-    }
-
-    // In case `Process::updated` is misused, `System::refresh_processes` might remove them
-    // so this test ensures that it doesn't happen.
-    #[test]
-    fn check_refresh_process_update() {
-        if !IS_SUPPORTED_SYSTEM {
-            return;
-        }
-        let mut s = System::new_all();
-        let total = s.processes().len() as isize;
-        s.refresh_processes();
-        let new_total = s.processes().len() as isize;
-        // There should be almost no difference in the processes count.
-        assert!(
-            (new_total - total).abs() <= 5,
-            "{} <= 5",
-            (new_total - total).abs()
-        );
-    }
-
-    #[test]
-    fn check_cpu_arch() {
-        assert_eq!(System::cpu_arch().is_some(), IS_SUPPORTED_SYSTEM);
-    }
-
-    // This test only exists to ensure that the `Display` and `Debug` traits are implemented on the
-    // `ProcessStatus` enum on all targets.
-    #[test]
-    fn check_display_impl_process_status() {
-        println!("{} {:?}", ProcessStatus::Parked, ProcessStatus::Idle);
-    }
-
-    // Ensure that the `Display` and `Debug` traits are implemented on the `MacAddr` struct
-    #[test]
-    fn check_display_impl_mac_address() {
-        println!(
-            "{} {:?}",
-            MacAddr([0x1, 0x2, 0x3, 0x4, 0x5, 0x6]),
-            MacAddr([0xa, 0xb, 0xc, 0xd, 0xe, 0xf])
-        );
-    }
-
-    #[test]
-    fn check_mac_address_is_unspecified_true() {
-        assert!(MacAddr::UNSPECIFIED.is_unspecified());
-        assert!(MacAddr([0; 6]).is_unspecified());
-    }
-
-    #[test]
-    fn check_mac_address_is_unspecified_false() {
-        assert!(!MacAddr([1, 2, 3, 4, 5, 6]).is_unspecified());
-    }
-
-    // Ensure that the `Display` and `Debug` traits are implemented on the `IpNetwork` struct
-    #[test]
-    fn check_display_impl_ip_network_ipv4() {
-        println!(
-            "{} {:?}",
-            IpNetwork {
-                addr: IpAddr::from(Ipv4Addr::new(1, 2, 3, 4)),
-                prefix: 3
-            },
-            IpNetwork {
-                addr: IpAddr::from(Ipv4Addr::new(255, 255, 255, 0)),
-                prefix: 21
-            }
-        );
-    }
-
-    #[test]
-    fn check_display_impl_ip_network_ipv6() {
-        println!(
-            "{} {:?}",
-            IpNetwork {
-                addr: IpAddr::from(Ipv6Addr::new(0xffff, 0xaabb, 00, 0, 0, 0x000c, 11, 21)),
-                prefix: 127
-            },
-            IpNetwork {
-                addr: IpAddr::from(Ipv6Addr::new(0xffcc, 0, 0, 0xffcc, 0, 0xffff, 0, 0xccaa)),
-                prefix: 120
-            }
-        )
-    }
-
-    #[test]
-    fn check_ip_networks() {
-        if !IS_SUPPORTED_SYSTEM {
-            return;
-        }
-        let networks = Networks::new_with_refreshed_list();
-        if networks.iter().any(|(_, n)| !n.ip_networks().is_empty()) {
-            return;
-        }
-        panic!("Networks should have at least one IP network ");
-    }
-
-    // This test exists to ensure that the `TryFrom<usize>` and `FromStr` traits are implemented
-    // on `Uid`, `Gid` and `Pid`.
-    #[allow(clippy::unnecessary_fallible_conversions)]
-    #[test]
-    fn check_uid_gid_from_impls() {
-        use std::convert::TryFrom;
-        use std::str::FromStr;
-
-        #[cfg(not(windows))]
-        {
-            assert!(crate::Uid::try_from(0usize).is_ok());
-            assert!(crate::Uid::from_str("0").is_ok());
-        }
-        #[cfg(windows)]
-        {
-            assert!(crate::Uid::from_str("S-1-5-18").is_ok()); // SECURITY_LOCAL_SYSTEM_RID
-            assert!(crate::Uid::from_str("0").is_err());
-        }
-
-        assert!(crate::Gid::try_from(0usize).is_ok());
-        assert!(crate::Gid::from_str("0").is_ok());
-
-        assert!(crate::Pid::try_from(0usize).is_ok());
-        // If it doesn't panic, it's fine.
-        let _ = crate::Pid::from(0);
-        assert!(crate::Pid::from_str("0").is_ok());
-    }
-
-    #[test]
-    fn check_groups() {
-        if !crate::IS_SUPPORTED_SYSTEM {
-            return;
-        }
-        assert!(!Groups::new_with_refreshed_list().is_empty());
     }
 }
